src/requirements.txt
absl-py==0.7.1
aniso8601==8.0.0
astor==0.8.0
astroid==2.3.3
attrs==19.3.0
audioread==2.1.8
backcall==0.1.0
certifi==2019.11.28
cffi==1.14.0
chardet==3.0.4
Click==7.0
clickclick==1.2.2
colorama==0.4.1
connexion==2.6.0
cycler==0.10.0
decorator==4.4.1
Flask==1.1.1
Flask-RESTful==0.3.8
gast==0.2.2
google-pasta==0.1.7
grpcio==1.22.0
gunicorn==20.0.4
h5py==2.9.0
idna==2.9
importlib-metadata==1.5.0
inflection==0.3.1
ipython==7.13.0
ipython-genutils==0.2.0
isort==4.3.21
itsdangerous==1.1.0
jedi==0.16.0
Jinja2==2.11.1
joblib==0.14.1
jsonify==0.5
jsonschema==3.2.0
Keras==2.3.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
kiwisolver==1.1.0
lazy-object-proxy==1.4.3
librosa==0.7.2
llvmlite==0.31.0
Markdown==3.1.1
MarkupSafe==1.1.1
matplotlib==3.1.3
mccabe==0.6.1
numba==0.48.0
numpy==1.16.4
openapi-spec-validator==0.2.8
pandas==1.0.1
parso==0.6.2
pickleshare==0.7.5
prompt-toolkit==3.0.3
protobuf==3.9.0
PyAudio==0.2.11
pycparser==2.19
Pygments==2.5.2
pylint==2.4.4
pyparsing==2.4.6
pyrsistent==0.15.7
python-dateutil==2.8.1
pytz==2019.3
PyYAML==5.3
requests==2.23.0
resampy==0.2.2
scikit-learn==0.22.1
scipy==1.4.1
seaborn==0.10.1
six==1.12.0
SoundFile==0.10.3.post1
tensorboard==1.14.0
tensorflow==1.14.0
tensorflow-estimator==1.14.0
termcolor==1.1.0
traitlets==4.3.3
typed-ast==1.4.0
urllib3==1.25.8
wcwidth==0.1.8
Werkzeug==0.15.5
wrapt==1.11.2
xlrd==1.2.0
XlsxWriter==1.2.8
zipp==3.1.0
